Refctor on styled components
