+++ src/pages/Products.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { ProductsGrid } from "../components/Products/Grid";
+import ProductsGrid from "../components/ProductsGrid";
 
 export const Products = () => {
   return <ProductsGrid>Products</ProductsGrid>;
